fix: fix mounting issues
