+++ packages/rainbowkit/src/components/ConnectButton/ConnectButton.tsx
@@ -39,7 +39,6 @@ export function ConnectButton({
       {({
         account,
         chain,
-        mounted,
         openAccountModal,
         openChainModal,
         openConnectModal,
@@ -47,19 +46,7 @@ export function ConnectButton({
         const unsupportedChain = chain?.unsupported ?? false;
 
         return (
-          <Box
-            display="flex"
-            gap="12"
-            transition="default"
-            {...(!mounted && {
-              'aria-hidden': true,
-              'style': {
-                opacity: 0,
-                pointerEvents: 'none',
-                userSelect: 'none',
-              },
-            })}
-          >
+          <Box display="flex" gap="12" transition="default">
             {account ? (
               <>
                 {chain && (chains.length > 1 || unsupportedChain) && (
