--- packages/rainbowkit/src/hooks/useIsMounted.ts
@@ -1,9 +1,0 @@
-import { useEffect, useState } from 'react';
-
-export const useIsMounted = () => {
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => setMounted(true), []);
-
-  return mounted;
-};
